component conversion from files into folders - separated react component files
