+++ src/Modules/Error/Info/strings.json
@@ -1,0 +1,5 @@
+{
+  "title": "Information Component",
+  "desc": "Brief info about the error page goes here",
+  "mainContent": "Error information section for the error module info component"
+}
