+++ src/Pages/Error/strings.json
@@ -1,0 +1,4 @@
+{
+  "title": "Error Page",
+  "description": "The page you're looking for could not be found. <br />Use our menu links to navigate to the list of our products"
+}
