+++ src/Pages/Error/styling.js
@@ -1,0 +1,20 @@
+export const cStyles = {
+  contentBox: {
+    margin: '8px 0px;',
+    padding: 16,
+  }
+}
+
+export const cProps = {
+  title: {
+    color: "textPrimary",
+    component: "h1",
+    variant: "h6",
+    gutterBottom: true
+  },
+  description: {
+    color: "textSecondary",
+    variant: "body1",
+    gutterBottom: true
+  }
+}
