+++ src/Pages/Home/strings.json
@@ -1,0 +1,4 @@
+{
+  "whatsNew": "Home Boilerplate - ReactJS/MUI/Webpack",
+  "newsAlert": "Desscription text for the home page goes here"
+}
