+++ src/index.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>Boilerplate with ReactJS/MUI/Webpack</title>
+  <title>Boilerplate with ReactJS/MUI/Webpack and State Management</title>
 </head>
 
 <body>
